Medoid does not need lines 50-56 since points in the Graph corelate only to one Filter

+++ Include/fun.hpp
@@ -19,6 +19,6 @@ void CreateKNNGraphBruteForce(Graph , int , const char *, const char *);
 
 
 /*Medoid Function*/
-int* FindMedoid(Graph *,int);
+int* FindMedoid(Graph *,float);
 
 #endif
